chore(backend): Extract signedIn/signedOut/interstitial to authStatus for testability
